Split failpoints file
